Create README
This project performs the simulation optical coherence tomography signals by assuming First-Born approximation for the scattered field.
The explanation of this OCT image formation model will be soon published.


%%%%%%%%%%%%%%%%%%  Journal for the pubblication: "Biomedical optics Express"    %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
Here the list of the files that have been used to generate all the figures of the related article: 

Figure (3) "Large C-scans with label 'TEST' "
	- Generated in the script "Large_OCT_Cscan_Taylor_new.m"   

Figure (4) "2 examples of application of the model: Simulation 1000 "small" OCT C-scans. This example is useful to show how this model can be used to train neuarl networks employing OCT data"
	- Generated in the script "Several_small_Cscans.m"  

Figure (5) "Compares the 3 approximations that are devolped into the model, including the two Taylor expansions and the Multispectral-regression"
	- Generated in the script "Comp_Rig_MSR_Tay_Tay2_new.m"  

Figure (6) "Test of the model on optical coherence elastography simulated data. (soft ellypse)"
	- Generated in the script "Comparison_OCE_rig_new.m"  

%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

%%%%%%%%%%%%%%%%%%     Matlab functions:    %%%%%%%%%%%%%%%%%%
"Series_Loaded_DW_Cscan_uniform_Taylor15" by using the fast OCT simulator model.  That function contains the second taylor expansion, which accelerates the simulation of the OCT C-scans of loaded data, which works only in axial direction and only for layers of uniform strains.

"Series_Loaded_DW_Cscan_uniform_Taylor16" It is a generalisation of "...15" which works for any distribution of axial strains, which are assigned to each scatterer of the simulation. On one hand this code is more general than "15", on the other hand the strains are interpolated a 3D-grid volume of space, in order to don't lose computational speed. 
    We suggest to use this function for any simulation involving only axial strains.

"Series_Loaded_DW_Cscan_uniform_Taylor_gen": It is a generalisation of  "...15"  which takes in input the scatterers positions of unloaded and loaded (one or more) C-scans and simulates the OCT signals for any motion of the scatterers, including lateral, axial strains and non linear scatterer displacements. In this case, however, the second Taylor expansion is not included, then multiple compression cases will be run separately. In that case, the scatterers transformation must be simulated by using a user-made compression simulator. 

"Comparison_15_16_gen.m" Shows how the three functions work and compares them for layers of uniform axial strains. The simulation runs 50 compressions of the same unloaded signal. "15-16" are faster than "gen" because they include the second taylor expansion 

"strain_UWA": Implementation of the WLS model for strain retrieval in OCE ( DOI:10.1364/BOE.5.002913)